requirements.txt
langgraph
langchain-community
langchain-text-splitters
langchain-anthropic
sentence-transformers
chromadb
fastapi
uvicorn
bs4
python-dotenv
